pyproject.toml
[project]
name = "lineart"
version = "0.0.0"
requires-python = ">=3.9,<3.13"
dependencies = [
    "torch",
    "diffusers",
    "transformers",
    "accelerate",
    "controlnet-aux",
    "pillow",
    "opencv-python",
    "scikit-image",
    "numpy",
    "requests",
    "huggingface-hub",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
gui = ["tkinterdnd2"]
svg = ["vtracer"]
gpu = [
    "xformers; platform_system == 'Linux' and platform_machine == 'x86_64'",
]

[project.scripts]
lineart-gui = "lineart.main:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E","F","W","I","UP","D","B","C90"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.deptry]

[tool.deptry.per_rule_ignores]
DEP001 = ["lineart"]
DEP002 = ["transformers", "accelerate", "xformers", "vtracer", "controlnet-aux"]
